Enhance: Move shamanistic rage into runes.go
